task_manager/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
python-jose[cryptography]
passlib
python-multipart